This is where I got tired.
On page load a list is created.  When clicked a task is editable.

On button click I'm not getting a successful getJSON call.  I've read
that I can pass get parameters as an optional argument, but I can't
seem to get it done.

Oh well, at least the back end is done.