Release candidate. There are things I can add, but I have to catch some sleep.
The main feature I will add is implicit "done"-ness. Right now
the user marks whether the task is done by using "true" or "false".
And that's terrible.  That will be remedied by setting isDone to
false on creation and then using a link to set it to complete.

And as I type this I realize that I'm leveraging the update
functionality to view a task, but that's not really in the ballpark
of best practices.  So once the week settles down I should fix that
as well.